# src/openapi.yaml
openapi: 3.0.1
info:
  title: Explore education statistics - public data API
  version: 1.0.0
  description: >
    This is a draft specification for the upcoming public data API for the EES platform.
  contact:
    name: Explore education statistics
    email: explore.statistics@education.gov.uk
    url: https://explore-education-statistics.service.gov.uk
servers:
  - url: https://ees-api-mock.ambitiousocean-cb084d07.uksouth.azurecontainerapps.io/
    description: Mock API
...
paths:
  /api/v1/publications:
    get:
      operationId: ListPublications
      summary: List publications
      description: Lists details about publications with data available for querying.
      tags:
        - Publication
      parameters:
        - name: page
          in: query
          description: The page of results to fetch.
          required: false
          schema:
            type: number
            minimum: 1
            format: int32
        - name: pageSize
          in: query
          description: The maximum number of results per page.
          required: false
          schema:
            type: number
            format: int32
            minimum: 0
            maximum: 40
            default: 20
        - name: search
          in: query
          description: A search term to find matching publications.
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedPublicationSummaryViewModel'
  /api/v1/publications/{publicationId}:
    get:
      operationId: GetPublication
      summary: Get a publication's details
      description: Gets a specific publication's details.
      tags:
        - Publication
      parameters:
        - name: publicationId
          in: path
          description: The ID of the publication.
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicationSummaryViewModel'
  /api/v1/publications/{publicationId}/data-sets:
    get:
      operationId: ListPublicationDataSets
      summary: List a publication's data sets
      description: Lists summary details of all the data sets related to a publication.
      tags:
        - Publication
      parameters:
        - name: publicationId
          in: path
          description: The ID of the publication.
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          description: The page of results to fetch.
          required: false
          schema:
            type: number
            minimum: 1
            format: int32
        - name: pageSize
          in: query
          description: The maximum number of results per page.
          required: false
          schema:
            type: number
            format: int32
            minimum: 0
            maximum: 20
            default: 10
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedDataSetsViewModel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/data-sets/{dataSetId}:
    get:
      operationId: GetDataSet
      summary: Get a data set's summary
      description: Gets a specific data set's summary details.
      tags:
        - Data Set
      parameters:
        - name: dataSetId
          in: path
          description: The ID of the data set.
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSetViewModel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/data-sets/{dataSetId}/query:
    get:
      operationId: QueryDataSetGet
      summary: Query a data set (GET)
      description: |-
        Query a data set using a `GET` request, returning the filtered results.

        Note that there is a `POST` variant of this endpoint which provides a more complete set
        of querying functionality. The `GET` variant is only recommended for initial exploratory 
        testing or simple queries that do not need advanced functionality.

        Unlike the `POST` variant, this endpoint does not allow condition clauses (`and`, `or`, `not`)
        and consequently cannot express more complex queries.
      tags:
        - Data Set
      parameters:
        - name: dataSetId
          in: path
          description: The ID of the data set.
          required: true
          schema:
            type: string
            format: uuid
        - name: dataSetVersion
          in: query
          description: The version of the data set to use e.g. 2.0, 1.1, etc.
          schema:
            type: string
        - name: filters
          in: query
          description: The filters to match results by.
          explode: false
          allowReserved: true
          schema:
            $ref: '#/components/schemas/DataSetQueryCriteriaFilters'
        - name: geographicLevels
          in: query
          description: The geographic levels to match results by.
          explode: false
          allowReserved: true
          schema:
            $ref: '#/components/schemas/DataSetQueryCriteriaGeographicLevels'
        - name: locations
          in: query
          description: The locations to match results by.
          explode: false
          allowReserved: true
          schema:
            $ref: '#/components/schemas/DataSetQueryCriteriaLocations'
        - name: timePeriods
          in: query
          description: The time periods to match results by.
          explode: false
          allowReserved: true
          schema:
            $ref: '#/components/schemas/DataSetQueryCriteriaTimePeriods'
        - name: indicators
          in: query
          description: The IDs or names of indicators to get values for.
          required: true
          explode: false
          allowReserved: true
          schema:
            type: array
            items:
              type: string
        - name: sort
          in: query
          description: |-
            List of sorts to apply to results. Sorts at the start of the list will be applied first.
            By default, results are sorted by time period in descending order.
          explode: false
          allowReserved: true
          schema:
            type: array
            minItems: 1
            items:
              $ref: '#/components/schemas/DataSetQuerySort'
        - name: debug
          in: query
          description: |-
            Enable debug mode. Results will be formatted with human-readable labels to assist identification.
          required: false
          schema:
            type: boolean
            default: false
        - name: page
          in: query
          description: The page of results to fetch.
          required: false
          schema:
            type: number
            format: int32
            minimum: 1
        - name: pageSize
          in: query
          description: The maximum number of results per page.
          required: false
          schema:
            type: number
            format: int32
            minimum: 0
            maximum: 10000
            default: 1000
        - name: Accept
          in: header
          description: |-
            The media type to format the response data in (defaults to `application/json`). Allowed options:

            - `application/json`
            - `text/csv`
          schema:
            type: string
            default: application/json
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSetResultsViewModel'
            text/csv:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      operationId: QueryDataSetPost
      summary: Query a data set (POST)
      description: |-
        Query a data set using a `POST` request, returning the filtered results.

        Note that for simpler queries or exploratory testing, there is also `GET` variant of this 
        endpoint which only handles a smaller subset of querying functionality. However, for most 
        use-cases, this endpoint is recommended as it provides the complete set of functionality.

        Unlike the `GET` variant, this endpoint allows condition clauses (`and`, `or`, `not`)
        and consequently can express more complex queries.
      tags:
        - Data Set
      parameters:
        - name: dataSetId
          in: path
          description: The ID of the data set.
          required: true
          schema:
            type: string
            format: uuid
        - name: dataSetVersion
          in: query
          description: The version of the data set to use e.g. 2.0, 1.1, etc.
          schema:
            type: string
        - name: debug
          in: query
          description: |-
            Enable debug mode. Results will be formatted with human-readable labels to assist identification.
          required: false
          schema:
            type: boolean
            default: false
        - name: page
          in: query
          description: The page of results to fetch.
          required: false
          schema:
            type: number
            format: int32
            minimum: 1
        - name: pageSize
          in: query
          description: The maximum number of results per page.
          required: false
          schema:
            type: number
            format: int32
            minimum: 0
            maximum: 10000
            default: 1000
        - name: Accept
          in: header
          description: |-
            The media type to format the response data in (defaults to `application/json`). Allowed options:

            - `application/json`
            - `text/csv`
          schema:
            type: string
            default: application/json
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataSetQuery'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSetResultsViewModel'
            text/csv:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/data-sets/{dataSetId}/file:
    get:
      operationId: GetDataSetFile
      summary: Get a data set file
      description: Get the entire data set in a specified file format.
      tags:
        - Data Set
      parameters:
        - name: dataSetId
          in: path
          description: The ID of the data set.
          required: true
          schema:
            type: string
            format: uuid
        - name: dataSetVersion
          in: query
          description: The version of the data set to use e.g. 2.0, 1.1, etc.
          schema:
            type: string
        - name: Accept
          in: header
          description: |-
            The media type to output the file in (defaults to `text/csv`). Allowed options:

            - `text/csv` - Outputs the data set as a compressed csv file (defaults to gzip).
            - `application/zip` - Outputs a zip file containing the data set, its metadata and any relevant files.
          schema:
            type: string
            default: text/csv
      responses:
        '200':
          description: Success. A zip containing the underlying data set file.
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/data-sets/{dataSetId}/meta:
    get:
      operationId: GetDataSetMeta
      summary: Get a data set's metadata
      description: Get the metadata about a data set. Use this to create data set queries.
      tags:
        - Data Set
      parameters:
        - name: dataSetId
          in: path
          description: The ID of the data set.
          required: true
          schema:
            type: string
            format: uuid
        - name: dataSetVersion
          in: query
          description: The version of the data set to use e.g. 2.0, 1.1, etc.
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSetMetaViewModel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/data-sets/{dataSetId}/meta/filters:
    get:
      operationId: GetDataSetMetaFilters
      summary: Get a data set's filter metadata
      description: Get the filter metadata about a data set. Use this to create data set queries.
      tags:
        - Data Set
      parameters:
        - name: dataSetId
          in: path
          description: The ID of the data set.
          required: true
          schema:
            type: string
            format: uuid
        - name: dataSetVersion
          in: query
          description: The version of the data set to use e.g. 2.0, 1.1, etc.
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSetMetaFiltersViewModel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/data-sets/{dataSetId}/meta/indicators:
    get:
      operationId: GetDataSetMetaIndicators
      summary: Get a data set's indicator metadata
      description: Get the indicator metadata about a data set. Use this to create data set queries.
      tags:
        - Data Set
      parameters:
        - name: dataSetId
          in: path
          description: The ID of the data set.
          required: true
          schema:
            type: string
            format: uuid
        - name: dataSetVersion
          in: query
          description: The version of the data set to use e.g. 2.0, 1.1, etc.
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSetMetaIndicatorsViewModel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/data-sets/{dataSetId}/meta/geographic:
    get:
      operationId: GetDataSetMetaGeographic
      summary: Get a data set's geographic metadata
      description: Get the geographic metadata about a data set. Use this to create data set queries.
      tags:
        - Data Set
      parameters:
        - name: dataSetId
          in: path
          description: The ID of the data set.
          required: true
          schema:
            type: string
            format: uuid
        - name: dataSetVersion
          in: query
          description: The version of the data set to use e.g. 2.0, 1.1, etc.
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSetMetaGeographicViewModel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/data-sets/{dataSetId}/meta/time-periods:
    get:
      operationId: GetDataSetMetaTimePeriods
      summary: Get a data set's time period metadata
      description: Get the time period metadata about a data set. Use this to create data set queries.
      tags:
        - Data Set
      parameters:
        - name: dataSetId
          in: path
          description: The ID of the data set.
          required: true
          schema:
            type: string
            format: uuid
        - name: dataSetVersion
          in: query
          description: The version of the data set to use e.g. 2.0, 1.1, etc.
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSetMetaTimePeriodsViewModel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/data-sets/{dataSetId}/versions:
    get:
      operationId: ListDataSetVersions
      summary: List a data set's versions
      description: List a data set's versions. Only provides summary information of each version.
      tags:
        - Data Set
      parameters:
        - name: dataSetId
          in: path
          description: The ID of the data set.
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          description: The page of results to fetch.
          required: false
          schema:
            type: number
            format: int32
            minimum: 1
        - name: pageSize
          in: query
          description: The maximum number of results per page.
          required: false
          schema:
            type: number
            format: int32
            minimum: 0
            maximum: 20
            default: 10
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedDataSetVersionsViewModel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/data-sets/{dataSetId}/versions/{dataSetVersion}:
    get:
      operationId: GetDataSetVersion
      summary: Get a data set version
      description: Get a data set version, including a full list of its changes.
      tags:
        - Data Set
      parameters:
        - name: dataSetId
          in: path
          description: The ID of the data set.
          required: true
          schema:
            type: string
            format: uuid
        - name: dataSetVersion
          in: path
          description: The data set version number e.g. 2.0, 1.1, etc.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSetVersionViewModel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/data-sets/{dataSetId}/versions/{dataSetVersion}/changes:
    get:
      operationId: GetDataSetVersionChanges
      summary: Get a data set version's changes
      description: Get a data set version's list of changes.
      tags:
        - Data Set
      parameters:
        - name: dataSetId
          in: path
          description: The ID of the data set.
          required: true
          schema:
            type: string
            format: uuid
        - name: dataSetVersion
          in: path
          description: The data set version number e.g. 2.0, 1.1, etc.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeSetViewModel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  responses:
    BadRequest:
      description: Bad Request. Usually indicates a validation error or a malformed request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorViewModel'
    NotFound:
      description: Not Found. Indicates that a specified resource could not be found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorViewModel'
  schemas:
...
    DataSetQueryCriteriaFilters:
      type: object
      description: The filter IDs to match results by.
      properties:
        eq:
          type: string
          description: Results must have a filter matching the given ID.
        notEq:
          type: string
          description: Results must not have a filter that matches the given ID.
        in:
          type: array
          description: Results must have at least one filter matching the given IDs.
          minItems: 1
          items:
            type: string
        notIn:
          type: array
          description: Results must not have any filters matching the given IDs.
          minItems: 1
          items:
            type: string
      additionalProperties: false
...
    PagedDataSetVersionsViewModel:
      type: object
      additionalProperties: false
      required:
        - paging
        - results
        - _links
      properties:
        paging:
          $ref: '#/components/schemas/PagingViewModel'
        results:
          type: array
          items:
            $ref: '#/components/schemas/DataSetVersionViewModel'
        _links:
          $ref: '#/components/schemas/LinksViewModel'
    PagedDataSetsViewModel:
      type: object
      additionalProperties: false
      required:
        - paging
        - results
        - _links
      properties:
        paging:
          $ref: '#/components/schemas/PagingViewModel'
        results:
          type: array
          items:
            $ref: '#/components/schemas/DataSetViewModel'
        _links:
          $ref: '#/components/schemas/LinksViewModel'
    PagedPublicationSummaryViewModel:
      type: object
      additionalProperties: false
      required:
        - paging
        - results
        - _links
      properties:
        paging:
          $ref: '#/components/schemas/PagingViewModel'
        results:
          type: array
          items:
            $ref: '#/components/schemas/PublicationSummaryViewModel'
        _links:
          $ref: '#/components/schemas/LinksViewModel'
...
    PublicationSummaryViewModel:
      type: object
      description: Provides summary information about a publication.
      required:
        - id
        - title
        - slug
        - summary
        - lastPublished
        - _links
      properties:
        id:
          type: string
          format: uuid
          description: The ID of the publication.
        title:
          type: string
          description: The title of the publication.
        slug:
          type: string
          description: The URL slug of the publication.
        summary:
          type: string
          description: The summary of the publication.
        lastPublished:
          type: string
          description: When the publication was last published.
          format: datetime
        _links:
          $ref: '#/components/schemas/LinksViewModel'
      additionalProperties: false
...